Update functional hog
This project is still in the works. The idea is that when my friend and I first created hog we were not using any sort of programming paradigm. We just wanted to get a playable hog game off the ground. After learning about functional programming, we thought it would be a good exercise to re-create hog but using only the functional programming paradigm. There is still much to learn however and we may have bit off more than we can chew. The project remains a work in progress however.